Adjust deployment for sepolia
