--- tasks/20230522-timelock-authorizer/input/goerli.json
@@ -1,152 +1,0 @@
-[
-  {
-    "taskId": "20210418-vault",
-    "contractName": "Vault",
-    "useAdaptor": false,
-    "signature": "manageUserBalance((uint8,address,uint256,address,address)[])",
-    "actionId": "0xeba777d811cd36c06d540d7ff2ed18ed042fd67bbf7c9afcf88c818c7ee6b498",
-    "grantee": "0x03f1ab8b19bce21eb06c364aec9e40322572a1e9",
-    "txHash": "0x79102fe6eb2ddddbd19801b99e1438635d14196806420150115ecf6f4e9ccb22"
-  },
-  {
-    "taskId": "20220325-mainnet-gauge-factory",
-    "contractName": "LiquidityGaugeV5",
-    "useAdaptor": true,
-    "target": "0xFFfFfFffFFfffFFfFFfFFFFFffFFFffffFfFFFfF",
-    "signature": "killGauge()",
-    "actionId": "0x38f12558087095621293389b35508390a03d17855b55b2394ce67b3c47d4463d",
-    "grantee": "0x3b1836c09fdae0a7c2848f99b2584aebd6a36876",
-    "txHash": "0xcf396873b2175efb35faf53f90818464e87c0b8ddcb874b5a4286202bbd080fe"
-  },
-  {
-    "taskId": "20220325-mainnet-gauge-factory",
-    "contractName": "LiquidityGaugeV5",
-    "useAdaptor": true,
-    "target": "0xFFfFfFffFFfffFFfFFfFFFFFffFFFffffFfFFFfF",
-    "signature": "add_reward(address,address)",
-    "actionId": "0x6b5e3750e9adcbd4a75428ed3fc6d2d3fc8a52866660ad3778d3f8c33108edff",
-    "grantee": "0x3babebfd684506a5b47701ee231a53427ad413ef",
-    "txHash": "0x112dc09e4819b11e8d768b7749b3c81ebd0e1cf042e5933dbba4e67ecdc14f42"
-  },
-  {
-    "taskId": "20220325-mainnet-gauge-factory",
-    "contractName": "LiquidityGaugeV5",
-    "useAdaptor": true,
-    "target": "0xFFfFfFffFFfffFFfFFfFFFFFffFFFffffFfFFFfF",
-    "signature": "killGauge()",
-    "actionId": "0x38f12558087095621293389b35508390a03d17855b55b2394ce67b3c47d4463d",
-    "grantee": "0xe0a171587b1cae546e069a943eda96916f5ee977",
-    "txHash": "0xff9653677fca74cbef95289e9fb4b65198378b8ac70447c515e058eda5b14dce"
-  },
-  {
-    "taskId": "20210418-vault",
-    "contractName": "Vault",
-    "useAdaptor": false,
-    "signature": "exitPool(bytes32,address,address,(address[],uint256[],bytes,bool))",
-    "actionId": "0xc149e88b59429ded7f601ab52ecd62331cac006ae07c16543439ed138dcb8d34",
-    "grantee": "0x03f1ab8b19bce21eb06c364aec9e40322572a1e9",
-    "txHash": "0x79102fe6eb2ddddbd19801b99e1438635d14196806420150115ecf6f4e9ccb22"
-  },
-  {
-    "taskId": "20210418-vault",
-    "contractName": "Vault",
-    "useAdaptor": false,
-    "signature": "joinPool(bytes32,address,address,(address[],uint256[],bytes,bool))",
-    "actionId": "0x78ad1b68d148c070372f8643c4648efbb63c6a8a338f3c24714868e791367653",
-    "grantee": "0x03f1ab8b19bce21eb06c364aec9e40322572a1e9",
-    "txHash": "0x79102fe6eb2ddddbd19801b99e1438635d14196806420150115ecf6f4e9ccb22"
-  },
-  {
-    "taskId": "20220325-mainnet-gauge-factory",
-    "contractName": "LiquidityGaugeV5",
-    "useAdaptor": true,
-    "target": "0xFFfFfFffFFfffFFfFFfFFFFFffFFFffffFfFFFfF",
-    "signature": "unkillGauge()",
-    "actionId": "0x5e7938ef4e6a59634c9800645277831eda1a30b5efd9ec55235154211f9e2941",
-    "grantee": "0x3b1836c09fdae0a7c2848f99b2584aebd6a36876",
-    "txHash": "0xa7591499e4495d65c277edd72c7e246a279c243b9c96182e4605136b1ed49a95"
-  },
-  {
-    "taskId": "20220420-smart-wallet-checker",
-    "contractName": "SmartWalletChecker",
-    "useAdaptor": false,
-    "signature": "denylistAddress(address)",
-    "actionId": "0xfa01325e2a9c77cda6b403aebfca7b8308d967df009e17ac6b42ba6d18088cfd",
-    "grantee": "0x3babebfd684506a5b47701ee231a53427ad413ef",
-    "txHash": "0xfe7eff0251d797eb3708b25c5af5a4f3c753ccf91842b8d46dddd6d6d21d9d12"
-  },
-  {
-    "taskId": "20220325-mainnet-gauge-factory",
-    "contractName": "LiquidityGaugeV5",
-    "useAdaptor": true,
-    "target": "0xFFfFfFffFFfffFFfFFfFFFFFffFFFffffFfFFFfF",
-    "signature": "set_reward_distributor(address,address)",
-    "actionId": "0x75f03d945d74639aa78784e5071604aa86c170a9f4c32bdf428231d560be70b7",
-    "grantee": "0xe0a171587b1cae546e069a943eda96916f5ee977",
-    "txHash": "0x3d2ac5880a7cd6d7ab2076c4400a6e043e5b81049f056492a3330e9c389ce6c5"
-  },
-  {
-    "taskId": "20220325-balancer-token-admin",
-    "contractName": "BalancerTokenAdmin",
-    "useAdaptor": false,
-    "signature": "mint(address,uint256)",
-    "actionId": "0x2315d78651468c46e4eb3cfca481a165fc94d355f1e38cb4e8c60fefdda8f86b",
-    "grantee": "0xdf0399539a72e2689b8b2dd53c3c2a0883879fdd",
-    "txHash": "0x3d2ac5880a7cd6d7ab2076c4400a6e043e5b81049f056492a3330e9c389ce6c5"
-  },
-  {
-    "taskId": "20210418-vault",
-    "contractName": "Vault",
-    "useAdaptor": false,
-    "signature": "batchSwap(uint8,(bytes32,uint256,uint256,uint256,bytes)[],address[],(address,bool,address,bool),int256[],uint256)",
-    "actionId": "0x1282ab709b2b70070f829c46bc36f76b32ad4989fecb2fcb09a1b3ce00bbfc30",
-    "grantee": "0x03f1ab8b19bce21eb06c364aec9e40322572a1e9",
-    "txHash": "0x79102fe6eb2ddddbd19801b99e1438635d14196806420150115ecf6f4e9ccb22"
-  },
-  {
-    "taskId": "20220420-smart-wallet-checker",
-    "contractName": "SmartWalletChecker",
-    "useAdaptor": false,
-    "signature": "allowlistAddress(address)",
-    "actionId": "0x552cf8a9722ea54b6fc81eef9df54cbbb0bdd80c1bdacf23d965b7bf7bd9f431",
-    "grantee": "0x171c0ff5943ce5f133130436a29bf61e26516003",
-    "txHash": "0x7787d3473cb9e8826a9495217f1d27105c4a86a847e1fd051de6cb05a2f99f43"
-  },
-  {
-    "taskId": "20220325-mainnet-gauge-factory",
-    "contractName": "LiquidityGaugeV5",
-    "useAdaptor": true,
-    "target": "0xFFfFfFffFFfffFFfFFfFFFFFffFFFffffFfFFFfF",
-    "signature": "add_reward(address,address)",
-    "actionId": "0x6b5e3750e9adcbd4a75428ed3fc6d2d3fc8a52866660ad3778d3f8c33108edff",
-    "grantee": "0xe0a171587b1cae546e069a943eda96916f5ee977",
-    "txHash": "0x3d2ac5880a7cd6d7ab2076c4400a6e043e5b81049f056492a3330e9c389ce6c5"
-  },
-  {
-    "taskId": "20220420-smart-wallet-checker",
-    "contractName": "SmartWalletChecker",
-    "useAdaptor": false,
-    "signature": "allowlistAddress(address)",
-    "actionId": "0x552cf8a9722ea54b6fc81eef9df54cbbb0bdd80c1bdacf23d965b7bf7bd9f431",
-    "grantee": "0x3babebfd684506a5b47701ee231a53427ad413ef",
-    "txHash": "0xfe7eff0251d797eb3708b25c5af5a4f3c753ccf91842b8d46dddd6d6d21d9d12"
-  },
-  {
-    "taskId": "20210418-vault",
-    "contractName": "Vault",
-    "useAdaptor": false,
-    "signature": "setRelayerApproval(address,address,bool)",
-    "actionId": "0x0014a06d322ff07fcc02b12f93eb77bb76e28cdee4fc0670b9dec98d24bbfec8",
-    "grantee": "0x03f1ab8b19bce21eb06c364aec9e40322572a1e9",
-    "txHash": "0x79102fe6eb2ddddbd19801b99e1438635d14196806420150115ecf6f4e9ccb22"
-  },
-  {
-    "taskId": "20210418-vault",
-    "contractName": "Vault",
-    "useAdaptor": false,
-    "signature": "swap((bytes32,uint8,address,address,uint256,bytes),(address,bool,address,bool),uint256,uint256)",
-    "actionId": "0x7b8a1d293670124924a0f532213753b89db10bde737249d4540e9a03657d1aff",
-    "grantee": "0x03f1ab8b19bce21eb06c364aec9e40322572a1e9",
-    "txHash": "0x79102fe6eb2ddddbd19801b99e1438635d14196806420150115ecf6f4e9ccb22"
-  }
-]
