--- tasks/20230522-timelock-authorizer/output/goerli.json
@@ -1,4 +1,0 @@
-{
-  "TimelockAuthorizerMigrator": "0x7159c4b96bfCfCA232355B55778e85Cf9E245192",
-  "TimelockAuthorizer": "0x70E26F740bAd6a2CDd762e2478496268882C21d4"
-}
